fix deleting attendance and creating attendance in terms of meetup timing

+++ graphql/resolvers/attendee.mutation.go
@@ -4,7 +4,6 @@ import (
     "context"
     "time"
 
-    "github.com/prprprus/scheduler"
     "github.com/secmohammed/meetups/mails"
     "github.com/secmohammed/meetups/middlewares"
     "github.com/secmohammed/meetups/models"
@@ -17,24 +16,25 @@ func (a *mutationResolver) CreateAttendance(ctx context.Context, input models.Cr
     if err != nil {
         return nil, errors.ErrRecordNotFound
     }
+    // disallow users to create attendance for meetups that's already after meetup end date.
+    if time.Now().After(meetup.EndDate) {
+        return nil, errors.ErrUnauthenticated
+    }
+
     attendee := &models.Attendee{
         Status:   input.Status.String(),
         MeetupID: input.MeetupID,
         UserID:   currentUser.ID,
     }
-    s, err := scheduler.NewScheduler(1000)
-    parsed, err := time.Parse("2006-01-02 15:04:05+02", meetup.StartDate)
-    if err != nil {
-        return nil, err
-    }
-    delayedTimeInSeconds := int(parsed.Add(time.Duration(-1) * time.Hour).Unix())
-    s.Delay().Second(delayedTimeInSeconds).Do(mails.SendReminderEmailToAttendee, currentUser, meetup)
+    delayedTimeInSeconds := int(meetup.StartDate.Add(time.Duration(-1) * time.Hour).Unix())
+    jobID := a.scheduler.Delay().Second(delayedTimeInSeconds).Do(mails.SendReminderEmailToAttendee, currentUser, meetup)
+    a.redisClient.Add(ctx, "job_id_for_meetup_"+meetup.ID, jobID)
     return a.AttendeesRepo.Create(attendee)
 }
 func (a *mutationResolver) DeleteAttendance(ctx context.Context, id string) (bool, error) {
     currentUser, _ := middlewares.GetCurrentUserFromContext(ctx)
     // TODO: to Delete attendance, it must be like sort of cancelation because a user mustn't delete his attendance of a meetup he attended.
-    attendee, err := a.AttendeesRepo.GetByID(id)
+    attendee, err := a.AttendeesRepo.GetByID(id, "Meetup")
     if err != nil {
         return false, errors.ErrRecordNotFound
     }
@@ -43,13 +43,21 @@ func (a *mutationResolver) DeleteAttendance(ctx context.Context, id string) (boo
     }
     // TODO: If the meetup hasn't been made yet, we must delete the reminder for the user.
     // to do so, we must store the schedule id that's made at redis.
+    if attendee.CreatedAt.After(attendee.Meetup.StartDate) {
+        return false, errors.ErrUnauthenticated
+    }
+    jobID, ok := a.redisClient.Get(ctx, "job_id_for_meetup_"+attendee.MeetupID)
+    if !ok {
+        return false, errors.ErrRecordNotFound
+    }
+    a.scheduler.CancelJob(jobID)
     return true, a.AttendeesRepo.Delete(attendee)
 
 }
 func (a *mutationResolver) UpdateAttendance(ctx context.Context, id string, status models.AttendanceStatus) (*models.Attendee, error) {
     currentUser, _ := middlewares.GetCurrentUserFromContext(ctx)
     // TODO: user can't update a status if the meeeting has passed away.
-    attendee, err := a.AttendeesRepo.GetByID(id)
+    attendee, err := a.AttendeesRepo.GetByID(id, "")
     if err != nil {
         return nil, errors.ErrRecordNotFound
     }
